Extract TextMessage, PeopleInSpace and CountryInformation
