+++ Program.cs
@@ -191,184 +191,6 @@ namespace UtilityBelt
 
     #region Choice Processors
 
-
-    #region Text Message
-
-    private static void TextMessage(IOptions<SecretsModel> options)
-    {
-      logger.LogInformation($"User Choice : {nameof(TextMessage)}");
-      Console.WriteLine();
-      Console.Write("Input number to receive message:");
-
-      string send2Number = Console.ReadLine().ToLower();
-
-      SmtpClient client = new SmtpClient("smtp.gmail.com", 587)
-      {
-        Credentials = new NetworkCredential(options.Value.Email, options.Value.EmailPassword),
-        EnableSsl = true
-      };
-
-      MailMessage message = new MailMessage();
-      message.From = new MailAddress(options.Value.Email);
-
-      var carrierType = MenuEnum<CarrierType>("Carrier");
-      var meta = EnumHelper.GetAttributeOfType<CarrierMetaAttribute>(carrierType);
-
-      message.To.Add(new MailAddress(send2Number + $"@{meta?.Domain}"));
-      message.Subject = "This is my subject";
-      message.Body = "This is the content";
-
-      try
-      {
-        client.Send(message);
-        Console.ForegroundColor = ConsoleColor.Cyan;
-        Console.WriteLine("Sent successfully");
-      }
-      catch
-      {
-        Console.ForegroundColor = ConsoleColor.Red;
-        Console.WriteLine("Text Message failed");
-      }
-
-      message.Dispose();
-      client.Dispose();
-    }
-
-    private static T MenuEnum<T>(string subject)
-            where T : struct, Enum
-    {
-      Console.WriteLine("");
-      Console.WriteLine($"Select the {subject}");
-
-      Array values = Enum.GetValues(typeof(T));
-      for (int i = 0; i < values.Length; i++)
-      {
-        var current = values.GetValue(i) as Enum;
-        var meta = EnumHelper.GetAttributeOfType<CarrierMetaAttribute>(current);
-        string description = meta?.Name ?? Enum.GetName(typeof(T), current);
-        Console.WriteLine($"{i + 1}) {description}");
-      }
-
-      Console.WriteLine("");
-
-      Console.Write("Your choice:");
-      string optionPicked = Console.ReadLine().ToLower();
-      logger.LogInformation($"User Choice : {optionPicked}");
-
-      bool isValid = Enum.TryParse(optionPicked, true, out T result);
-
-      if (!isValid || !Enum.IsDefined(typeof(T), result))
-      {
-        logger.LogWarning($"Invalid User Choice : {optionPicked}");
-        Console.WriteLine("Please select a valid option");
-        return MenuEnum<T>(subject);
-      }
-
-      return result;
-    }
-
-    #endregion Text Message
-
-    #region Cat facts
-
-    private static void CatFact()
-    {
-      logger.LogInformation($"User Choice : {nameof(CatFact)}");
-      string content = string.Empty;
-      string catUrl = "https://cat-fact.herokuapp.com/facts/random";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(catUrl);
-      }
-      CatFactModel catFact = JsonSerializer.Deserialize<CatFactModel>(content);
-      Console.WriteLine();
-      if (catFact.Status != null && catFact.Status.Verified)
-        Console.ForegroundColor = ConsoleColor.Yellow;
-      else
-        Console.ForegroundColor = ConsoleColor.Red;
-      Console.WriteLine(catFact.Text);
-      Console.WriteLine();
-    }
-
-    #endregion Cat facts
-
-    #region People in space
-
-    private static void Space()
-    {
-      logger.LogInformation($"User Choice : {nameof(Space)}");
-      string content = string.Empty;
-      string spacePeopleUrl = "http://api.open-notify.org/astros.json";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(spacePeopleUrl);
-      }
-      SpacePersonModel spacePeopleFact = JsonSerializer.Deserialize<SpacePersonModel>(content);
-      Console.WriteLine();
-      Console.ForegroundColor = ConsoleColor.Yellow;
-      Console.WriteLine("There are " + spacePeopleFact.People.Count + " in space right now!");
-      foreach (SpacePerson spacePerson in spacePeopleFact.People)
-      {
-        Console.WriteLine(spacePerson.Name + " is in " + spacePerson.Craft);
-      }
-      Console.WriteLine();
-    }
-
-    #endregion People in space
-
-    #region Country information
-
-    private static void CountryInformation()
-    {
-      logger.LogInformation($"User Choice : {nameof(CountryInformation)}");
-      string content = string.Empty;
-      Console.WriteLine("");
-      Console.WriteLine("Please enter a country name:");
-
-      string countryName = Console.ReadLine().ToLower();
-      string url = $"https://restcountries.eu/rest/v2/name/{countryName}";
-
-      try
-      {
-        using (var wc = new WebClient())
-        {
-          content = wc.DownloadString(url);
-        }
-
-        var countryInformationResult = JsonSerializer.Deserialize<List<CountryInformation>>(content);
-
-        foreach (var item in countryInformationResult)
-        {
-          Console.WriteLine("===============================================");
-          Console.WriteLine($"Country Name: {item.Name}");
-          Console.WriteLine($"Capital: {item.Capital}");
-          Console.WriteLine($"Region: {item.Region}");
-          Console.WriteLine($"Population: {item.Population.ToString("N1")}");
-          Console.WriteLine($"Area: {item.Area.ToString("N1")} km²");
-          Console.WriteLine("Currencies");
-          foreach (var moneda in item.Currencies)
-          {
-            Console.WriteLine($"*Code:\t\t{moneda.Code}");
-            Console.WriteLine($"*Name:\t\t{moneda.Name}");
-            Console.WriteLine($"*Symbol:\t{moneda.Symbol}");
-          }
-          Console.WriteLine("Languages");
-          foreach (var language in item.Languages)
-          {
-            Console.WriteLine($"* Name:\t\t{language.Name} / {language.NativeName}");
-          }
-          Console.WriteLine("===============================================");
-        }
-      }
-      catch (WebException)
-      {
-        logger.LogWarning("Country not found");
-        Console.WriteLine("Country not found");
-      }
-    }
-
-    #endregion Country information
-
     #region Discord Sender
 
     private static void DiscordWebhook(IOptions<SecretsModel> options)
